+++ src/api/admin/posts.ts
@@ -1,4 +1,5 @@
-import { apiClient, type ApiResponse } from '@/lib/api';
+import { apiClient } from '@/lib/api';
+import type { ApiResponse } from '@/types/api/client';
 
 import type {
   ActivityType,
@@ -113,10 +114,6 @@ export const postsApi = {
   // 게시글 상세 조회
   async getPostDetail(params: GetPostDetailParams): Promise<ApiResponse<PostDetailResponse>> {
     try {
-      console.log('getPostDetail params:', params);
-      console.log('postId type:', typeof params.postId);
-      console.log('postId value:', params.postId);
-
       // postId가 이미 숫자인지 확인
       let postId: number;
       if (typeof params.postId === 'number') {
@@ -124,8 +121,6 @@ export const postsApi = {
       } else {
         postId = parseInt(params.postId, 10);
       }
-
-      console.log('final postId:', postId);
 
       if (isNaN(postId) || postId <= 0) {
         throw new Error(`Invalid post ID: ${params.postId}`);
@@ -144,10 +139,6 @@ export const postsApi = {
     params: GetPostDetailParams
   ): Promise<ApiResponse<CalendarPostDetailResponse>> {
     try {
-      console.log('getCalendarPostDetail params:', params);
-      console.log('postId type:', typeof params.postId);
-      console.log('postId value:', params.postId);
-
       // postId가 이미 숫자인지 확인
       let postId: number;
       if (typeof params.postId === 'number') {
@@ -155,8 +146,6 @@ export const postsApi = {
       } else {
         postId = parseInt(params.postId, 10);
       }
-
-      console.log('final postId:', postId);
 
       if (isNaN(postId) || postId <= 0) {
         throw new Error(`Invalid post ID: ${params.postId}`);
